Continuing developing contexmenuevent(). Also moved ImageProcessor() class to ImageProcessor.py
